+++ HISTORY.rst
@@ -10,6 +10,8 @@ History
 * Fixed bug where 'void_ratio' method for unit weight in BI2014 triggering added dry weight twice
 * Added support for loading cpt files that do not have the u2 column
 * Updated factor of safety colors for plotting, added color map and colors accessible as static values
+* Added npts to bi2014 object
+* Added colors and color map for I_c soil values
 
 0.5.3 (2019-04-08)
 -------------------